// src/modules/shift/enums/shift.enum.ts
export enum ENUM_SHIFT_ASSIGNMENT_STATUS {
    SCHEDULED = 'scheduled',
    CONFIRMED = 'confirmed',
    SWAPPED = 'swapped',
    CANCELLED = 'cancelled',
}

export enum ENUM_SHIFT_SWAP_STATUS {
    PENDING = 'pending',
    ACCEPTED = 'accepted',
    REJECTED = 'rejected',
    ADMIN_APPROVED = 'admin_approved',
    ADMIN_REJECTED = 'admin_rejected',
}